Require methods *everywhere*
I would love to know why this package is considered optional to load
when basically *nothing works properly without it*.  As of 3.1.0,
Rscript seems to have become even less useful in its basic form,
generating "could not find function "is"" errors everywhere (confusingly
things still run).  Inredible that ommiting basic features of the language
is considered a sensible thing to do for the only front end.